requirements.txt
beautifulsoup4
lxml
pandas
openai
tqdm
py-readability-metrics